+++ fractals/src/main/java/me/catmousedog/fractals/fractals/types/normalized/NormalizedJulia.java
@@ -1,34 +1,24 @@
 package me.catmousedog.fractals.fractals.types.normalized;
 
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseMotionListener;
 import java.util.Properties;
 
 import me.catmousedog.fractals.fractals.Fractal;
-import me.catmousedog.fractals.fractals.filters.Filter;
-import me.catmousedog.fractals.fractals.filters.LogPeriodicFilter;
+import me.catmousedog.fractals.fractals.abstract_fractals.Julia;
 import me.catmousedog.fractals.main.Settings;
-import me.catmousedog.fractals.ui.components.Item;
-import me.catmousedog.fractals.ui.components.concrete.CheckBox;
-import me.catmousedog.fractals.ui.components.concrete.Padding;
-import me.catmousedog.fractals.ui.components.concrete.TextFieldDouble;
 
 /**
  * Number = Integer
  */
-public final class NormalizedJulia extends Fractal {
+public final class NormalizedJulia extends Julia {
 
 	private int offset;
-	private double jx, jy;
 
 	public NormalizedJulia(Settings settings) {
 		super(settings);
 	}
 
 	private NormalizedJulia(Settings settings, Fractal fractal, double jx, double jy, int offset) {
-		super(settings, fractal);
-		this.jx = jx;
-		this.jy = jy;
+		super(settings, fractal, jx, jy);
 		this.offset = offset;
 	}
 
@@ -58,20 +48,6 @@ public final class NormalizedJulia extends Fractal {
 	}
 
 	@Override
-	public void save() {
-		jx = jxjtf.saveAndGet();
-		jy = jyjtf.saveAndGet();
-		super.save();
-	}
-
-	@Override
-	public void update() {
-		jxjtf.setData(jx);
-		jyjtf.setData(jy);
-		super.update();
-	}
-
-	@Override
 	public String informalName() {
 		return "Normalized Julia Set";
 	}
@@ -86,49 +62,6 @@ public final class NormalizedJulia extends Fractal {
 		return "<html>The second order julia set (<i>z²+c<i/>) generated using an normalized iteration counts."
 				+ "<br>This allows for deep zooms but is slower."
 				+ "<br>Dragging the mouse will change the fixed point used to generate this set.</html>";
-	}
-
-	private TextFieldDouble jxjtf;
-
-	private TextFieldDouble jyjtf;
-
-	private CheckBox mousejcx;
-
-	@Override
-	protected void initFractal() {
-		filters = new Filter[] { new LogPeriodicFilter(this) };
-		filter = filters[0];
-		mouse = new MouseMotionListener() {
-			@Override
-			public void mouseMoved(MouseEvent e) {
-			}
-
-			@Override
-			public void mouseDragged(MouseEvent e) {
-				if (mousejcx.saveAndGet()) {
-					double[] t = transform.apply(e.getX(), e.getY());
-					jx = t[0];
-					jy = t[1];
-					update();
-					if (settings.isRender_on_changes())
-						jpi.renderNow();
-				}
-			}
-		};
-
-		Padding p5 = new Padding(5);
-
-		mousejcx = new CheckBox.Builder("mouse listener")
-				.setTip("<html>If enabled, the user can click and drag the mouse along the canvas<br>"
-						+ " to change the fixed julia point.</html>")
-				.build();
-
-		jxjtf = new TextFieldDouble.Builder().setLabel("Jx").setTip("The fixed julia point's x-coordinate")
-				.setDefault(jx).build();
-		jyjtf = new TextFieldDouble.Builder().setLabel("Jy").setTip("The fixed julia point's y-coordinate")
-				.setDefault(jy).build();
-
-		items = new Item[] { jxjtf, p5, jyjtf, p5, mousejcx };
 	}
 
 	@Override